Refactor aws2-quarkus-client-ddb(s3) by copying files (not duplicating) #3236
