+++ integration-test-groups/aws2-quarkus-client/aws2-ddb/src/test/resources/META-INF/services/org.apache.camel.quarkus.test.support.aws2.Aws2TestEnvCustomizer
@@ -1,1 +1,1 @@
-org.apache.camel.quarkus.component.aws2.ddb.it.Aws2DdbTestEnvCustomizer
+org.apache.camel.quarkus.component.aws2.ddb.it.Aws2DdbQuarkusClientTestEnvCustomizer
