+++ integration-tests/aws2-quarkus-client-grouped/pom.xml
@@ -178,6 +178,21 @@
                         </configuration>
                     </execution>
                     <execution>
+                        <id>copy-client-test-sources-ddb</id>
+                        <goals>
+                            <goal>execute</goal>
+                        </goals>
+                        <phase>generate-sources</phase>
+                        <configuration>
+                            <source>file:///${maven.multiModuleProjectDirectory}/tooling/scripts/copy-tests.groovy</source>
+                            <properties>
+                                <copy-tests.source.dir>${maven.multiModuleProjectDirectory}/integration-test-groups/aws2/aws2-ddb</copy-tests.source.dir>
+                                <copy-tests.dest.module.dir>${project.basedir}</copy-tests.dest.module.dir>
+                                <copy-tests.excludes>**/*TestEnvCustomizer,**/*application.properties,**/*Stream*</copy-tests.excludes>
+                            </properties>
+                        </configuration>
+                    </execution>
+                    <execution>
                         <id>copy-client-test-sources-s3</id>
                         <goals>
                             <goal>execute</goal>
